[Rename] addressed review feedbacks
