+++ src/plugins/vis_type_table/public/agg_table/agg_table.test.js
@@ -26,7 +26,7 @@ import { round } from 'lodash';
 
 import { getFieldFormatsRegistry } from '../../../data/public/test_utils';
 import { coreMock } from '../../../../core/public/mocks';
-import { initAngularBootstrap } from '../../../opensearch_dashboardslegacy/public';
+import { initAngularBootstrap } from '../../../opensearch_dashboards_legacy/public';
 import { setUiSettings } from '../../../data/public/services';
 import { UI_SETTINGS } from '../../../data/public/';
 import { CSV_SEPARATOR_SETTING, CSV_QUOTE_VALUES_SETTING } from '../../../share/public';
