+++ src/plugins/vis_type_table/public/paginated_table/paginated_table.test.ts
@@ -27,7 +27,7 @@ import { getAngularModule } from '../get_inner_angular';
 import { initTableVisLegacyModule } from '../table_vis_legacy_module';
 import { coreMock } from '../../../../core/public/mocks';
 
-jest.mock('../../../opensearch_dashboardslegacy/public/angular/angular_config', () => ({
+jest.mock('../../../opensearch_dashboards_legacy/public/angular/angular_config', () => ({
   configureAppAngularModule: () => {},
 }));
